feat: made ranomdizer a reusable component
